Start a basic functional test framework.
We use a standard TAP functional test format with `tNNN*.js` test
scripts under a `t/` subdirectory. These can be run directly from
the project root with `node t/t001.js` or similar, and later the
build system will be extended to find and run tests automatically.

The `t/tlib.js` file is a library of handy support functions for
this, including bringing it various test tools (tape, tape-spawn,
etc.) and functions to read in data files that store expected output
and whatnot. tlib itself is tested with `t/t001-tlib.js`.

The first test of the non-test-framework code is `t010-version.js`,
which proves we can run the `gapps` binary.