Rename project to gas-tools.
We're effectively forking the project here as our development path (and
even some current features, e.g., the "download" command) are moving in
a different direction from the one in which danthareja wants to go.

The command-line tool is renamed to `gas`, to avoid conflict with those
who have projects that use the `gapps` tool.

The local configuration file for the GAS projects is renamed to avoid
confusion and because we'll soon be changing the format, but we will
continue to use the `$HOME/.gapps` file for Google Drive authentication
information until we expand the format of that to handle multiple
accounts.

We did a quick pass over the README to fix up the most obvious things
related to the rename, but there's much more to be done there.