t020-init: No need to use promises, since tape runs tests synchronously.
Tape (and at least some things that use it, such as tape-spawn) ensures
that tests run one-at-a-time and in order. Thus, we don't need to use
promises to enforce this; we can just put the tests in a particular file
in dependency order.

This actually isn't terribly pretty or comfortable, but it's the best
we've got for now.